(registration-react) refactor catalogs-page with connected-catalogs-page and fetch apis
