+++ applications/registration-react/src/pages/catalogs-page/catalogs-page.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { CardGroup } from 'reactstrap';
 import _ from 'lodash';
@@ -7,14 +6,18 @@ import { FeatureToggle } from 'react-feature-toggles';
 
 import localization from '../../utils/localization';
 import { FEATURES } from '../../app-protected-route/features';
-import { fetchCatalogsIfNeeded } from '../../redux/modules/catalogs';
-import { fetchDatasetsIfNeeded } from '../../redux/modules/datasets';
 import { Catalog } from './catalogs/catalogs.component';
 import getTranslateText from '../../utils/translateText';
 import './catalogs-page.scss';
 
 const renderCatalogs = props => {
-  const { catalogItems, datasets, apis, fetchDatasetsIfNeeded } = props;
+  const {
+    catalogItems,
+    datasets,
+    apis,
+    fetchDatasetsIfNeeded,
+    fetchApisIfNeeded
+  } = props;
 
   if (!catalogItems) {
     return null;
@@ -44,6 +47,7 @@ const renderCatalogs = props => {
               <Catalog
                 key={`apis-${item.id}`}
                 catalogId={item.id}
+                fetchItems={fetchApisIfNeeded}
                 type="apis"
                 items={apis}
               />
@@ -100,27 +104,3 @@ RegCatalogs.propTypes = {
   isFetching: PropTypes.bool,
   fetchCatalogsIfNeeded: PropTypes.func
 };
-
-function mapStateToProps({ catalogs, datasets, apis }) {
-  const { catalogItems, isFetching } = catalogs || {
-    catalogItems: null
-  };
-
-  return {
-    catalogItems,
-    datasets,
-    apis,
-    isFetching
-  };
-}
-
-const mapDispatchToProps = dispatch => ({
-  fetchCatalogsIfNeeded: catalogsURL =>
-    dispatch(fetchCatalogsIfNeeded(catalogsURL)),
-  fetchDatasetsIfNeeded: catalogId => dispatch(fetchDatasetsIfNeeded(catalogId))
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(RegCatalogs);
